pyproject.toml
[project]
name = "little-python-toy-demos"
version = "0.1.0"
description = "Small Python toy demos (snake, voxels, etc.)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "noise>=1.2.2",
    "numpy>=2.2.2",
    "pygame>=2.6.1",
    "pyopengl>=3.1.9",
]

[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lptd"]
